+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes will be documented in this file.
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
-## [Unreleased]
+## [1.1.0] - 2021-06-25
 ### Added
 - Add PHP 7.2 compatibility
 - Add `EnvelopeTag.width`